Create Transitions with PERMUTATIONS
Function "transition" used to determine whether the behavior sequence follows a non random pattern using permutations. This method created nsim random transitions among behavioral units using the observed sequences, then determined the proportion of random transitions where the values were equal or higher than the value of the observed transition. This proportion represents the probability that the transition observed happened at random. This method gives us an intuitive proxy of non-random transitions without following any specific distribution of errors.
To run the function you need a table with two columns (data): a column with the identity of the individual (or pair) which performed the behaviors and another with the sequence temporarily ordered of the behaviors performed. You also have to define the number of permutations you want (n.sim) and if you want to output an observed matrix of transitions by individual (ind.matrix).